Extract windows-specific code
