+++ windows/Cargo.toml
@@ -25,7 +25,7 @@ strip = true # Remove debug symbols
 
 [dependencies]
 directories = "4"
-glow = "0.11.2"
+glow = "0.12.2"
 log = { version = "0.4", features = ["serde"] }
 log-panics = { version = "2", features = ["with-backtrace"]}
 raw-window-handle = "0.5"
@@ -40,12 +40,6 @@ git = "https://github.com/sandydoo/winit"
 branch = "flux"
 default-features = false
 
-# [target.'cfg(not(windows))'.dependencies.sdl2]
-# version = "0.35.2"
-# default-features = false
-# features = [ "use-pkgconfig", "raw-window-handle" ]
-
-# [target.'cfg(windows)'.dependencies.sdl2]
 [dependencies.sdl2]
 version = "0.35.2"
 default-features = false
@@ -80,16 +74,22 @@ branch = "main"
 version =  "^0.48.0"
 features = [
   "Win32_Foundation",
+  "Win32_Graphics_Direct3D",
+  "Win32_Graphics_Direct3D11",
   "Win32_Graphics_Dwm",
+  "Win32_Graphics_Dxgi",
+  "Win32_Graphics_Dxgi_Common",
   "Win32_Graphics_Gdi",
+  "Win32_Graphics_OpenGL",
   "Win32_System_Com",
   "Win32_System_LibraryLoader",
+  "Win32_System_Threading",
   "Win32_UI_HiDpi",
   "Win32_UI_Shell",
   "Win32_UI_WindowsAndMessaging"
 ]
 
-[build-dependencies]
+[target.'cfg(windows)'.build-dependencies]
 winres = "0.1.12"
 
 [patch.crates-io]
