+++ windows/src/platform/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(windows)]
+pub mod windows;
