+++ windows/src/platform/windows/dpi_awareness.rs
@@ -1,0 +1,26 @@
+// Specifying DPI awareness in the app manifest does not apply when running in a
+// preview window.
+pub fn set_dpi_awareness() -> Result<(), String> {
+    use windows::Win32::Foundation::E_INVALIDARG;
+    use windows::Win32::UI::HiDpi::{
+        GetProcessDpiAwareness, SetProcessDpiAwareness, PROCESS_PER_MONITOR_DPI_AWARE,
+        PROCESS_SYSTEM_DPI_AWARE,
+    };
+
+    if let Err(err) = unsafe { SetProcessDpiAwareness(PROCESS_PER_MONITOR_DPI_AWARE) } {
+        match err.code() {
+            E_INVALIDARG => return Err("Can’t enable support for high-resolution screens.".to_string()),
+            // The app manifest settings, if applied, trigger this path.
+            _ => {
+                return match unsafe { GetProcessDpiAwareness(None) } {
+                    Ok(awareness)
+                        if awareness == PROCESS_PER_MONITOR_DPI_AWARE
+                        || awareness == PROCESS_SYSTEM_DPI_AWARE => Ok(()),
+                    _ => Err("Can’t enable support for high-resolution screens. The setting has been modified and set to an unsupported value.".to_string()),
+                }
+            }
+        }
+    }
+
+    Ok(())
+}
